Run each markdown file as its own test
Discover markdown files with datatest-stable so files run in parallel, failures name the file, and database state cannot leak between validation files. Each validation file gets one database, which its scenarios still share in source order.
