+++ crates/djls-semantic/Cargo.toml
@@ -24,8 +24,14 @@ which = { workspace = true }
 
 [dev-dependencies]
 djls-testing = { workspace = true }
+
+datatest-stable = { workspace = true }
 insta = { workspace = true }
 tempfile = { workspace = true }
+
+[[test]]
+name = "mdtest"
+harness = false
 
 [lints]
 workspace = true
