+++ crates/djls-testing/src/lib.rs
@@ -48,8 +48,8 @@ pub use fixtures::validation_db;
 pub use mdtest::Scenario;
 pub use mdtest::ScenarioFile;
 pub use mdtest::ScenarioFileKind;
-pub use mdtest::run_suite;
-pub use mdtest::run_suite_with;
+pub use mdtest::run_file;
+pub use mdtest::run_file_with;
 pub use settings::ProjectSettings;
 pub use vendor::VendorSpecFixturesOptions;
 pub use vendor::vendor_spec_fixtures;
